Drop Haml helpers
