--- test/results/helpers.xhtml
@@ -1,72 +1,0 @@
-&amp;&amp;&amp;&amp;&amp;&amp;&amp;&amp;&amp;&amp;&amp;
-<div>
-<p class='title'>Title</p>
-<p class='text'>
-Woah this is really crazy
-I mean wow,
-man.
-</p>
-</div>
-
-<div>
-<p class='title'>Title</p>
-<p class='text'>
-Woah this is really crazy
-I mean wow,
-man.
-</p>
-</div>
-
-<div>
-<p class='title'>Title</p>
-<p class='text'>
-Woah this is really crazy
-I mean wow,
-man.
-</p>
-</div>
-
-<p>foo</p>
-<p>reeeeeeeeeeeeeeeeeeeeeeeeeeeeeeally loooooooooooooooooong</p>
-<div class='woah'>
-<div id='funky'>
-<div>
-<h1>Big!</h1>
-<p>Small</p>
-<!-- Invisible -->
-</div>
-<div class='dilly'>
-<p>foo</p>
-<h1>bar</h1>
-</div>
-</div>
-(<strong>parentheses!</strong>)
-</div>
-*<span class='small'>Not really</span>
-click
-<a href='thing'>here</a>.
-<p>baz</p>
-<p>boom</p>
-foo
-<li><a href='http://www.google.com'>google</a></li><p>
-foo
-<div>
-bar
-</div>
-boom
-baz
-boom, again
-</p>
-<table>
-  <tr>
-    <td class='cell'>
-      <strong>strong!</strong>
-      data
-    </td>
-    <td>
-      more_data
-    </td>
-  </tr>
-</table>
-<hr />
-<div></div>
