--- test/results/nuke_outer_whitespace.xhtml
@@ -1,140 +1,0 @@
-<p>
-<p><q>
-Foo
-</q></p>
-</p>
-<p>
-<p><q a='2'>
-Foo
-</q></p>
-</p>
-<p>
-<p><q>Foo</q></p>
-</p>
-<p>
-<p><q a='2'>Foo</q></p>
-</p>
-<p>
-<p><q>
-Foo
-</q></p>
-</p>
-<p>
-<p><q a='2'>
-Foo
-</q></p>
-</p>
-<p>
-<p><q>Foo</q></p>
-</p>
-<p>
-<p><q a='2'>Foo</q></p>
-</p>
-<p>
-<p><q>
-Foo
-Bar
-</q></p>
-</p>
-<p>
-<p><q a='2'>
-Foo
-Bar
-</q></p>
-</p>
-<p>
-<p><q>Foo
-Bar</q></p>
-</p>
-<p>
-<p><q a='2'>Foo
-Bar</q></p>
-</p>
-<p>
-<p>
-foo<q>
-Foo
-</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q a='2'>
-Foo
-</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q>Foo</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q a='2'>Foo</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q>
-Foo
-</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q a='2'>
-Foo
-</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q>Foo</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q a='2'>Foo</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q>
-Foo
-Bar
-</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q a='2'>
-Foo
-Bar
-</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q>Foo
-Bar</q>bar
-</p>
-</p>
-<p>
-<p>
-foo<q a='2'>Foo
-Bar</q>bar
-</p>
-</p>
-<p>
-<p><q></q></p>
-</p>
-<p>
-<p><q /></p>
-</p>
-<p>
-<p><q a='2'></q></p>
-</p>
-<p>
-<p><q a='2' /></p>
-</p>
