--- test/results/partial_layout.xhtml
@@ -1,6 +1,0 @@
-<h1>Partial layout used with for block:</h1>
-<div class='partial-layout'>
-<h2>This is inside a partial layout</h2>
-<p>Some content within a layout</p>
-
-</div>
