--- test/results/whitespace_handling.xhtml
@@ -1,91 +1,0 @@
-<div id='whitespace_test'>
-<div class='text_area_test_area'>
-<textarea>Oneline</textarea>
-</div>
-<textarea>BLAH
-</textarea>
-
-<div class='text_area_test_area'>
-<textarea>Two&#x000A;lines</textarea>
-</div>
-<textarea>BLAH
-</textarea>
-
-<div class='text_area_test_area'>
-<textarea>Oneline</textarea>
-</div>
-<textarea>BLAH</textarea>
-
-<div class='text_area_test_area'>
-<textarea>Two&#x000A;lines</textarea>
-</div>
-<textarea>BLAH</textarea>
-
-<div id='flattened'><div class='text_area_test_area'>
-<textarea>Two&#x000A;lines</textarea>
-</div>
-<textarea>BLAH</textarea>
-</div>
-</div>
-<div class='hithere'>
-Foo bar
-<pre>foo bar</pre>
-<pre>foo&#x000A;bar</pre>
-<p><pre>foo&#x000A;bar</pre></p>
-<p>foo
-bar</p>
-</div>
-<div class='foo'>
-13
-<textarea>
-a
-</textarea>
-<textarea>
-a</textarea></div>
-<div id='whitespace_test'>
-<div class='text_area_test_area'>
-<textarea>Oneline</textarea>
-</div>
-<textarea>BLAH
-</textarea>
-
-<div class='text_area_test_area'>
-<textarea>Two&#x000A;lines</textarea>
-</div>
-<textarea>BLAH
-</textarea>
-
-<div class='text_area_test_area'>
-<textarea>Oneline</textarea>
-</div>
-<textarea>BLAH</textarea>
-
-<div class='text_area_test_area'>
-<textarea>Two&#x000A;lines</textarea>
-</div>
-<textarea>BLAH</textarea>
-
-<div id='flattened'><div class='text_area_test_area'>
-<textarea>Two&#x000A;lines</textarea>
-</div>
-<textarea>BLAH</textarea>
-</div>
-</div>
-<div class='hithere'>
-Foo bar
-<pre>foo bar</pre>
-<pre>foo&#x000A;bar</pre>
-<p><pre>foo&#x000A;bar</pre></p>
-<p>foo
-bar</p>
-<pre>                                                 ___&#x000A;                                              ,o88888&#x000A;                                           ,o8888888'&#x000A;                     ,:o:o:oooo.        ,8O88Pd8888"&#x000A;                 ,.::.::o:ooooOoOoO. ,oO8O8Pd888'"&#x000A;               ,.:.::o:ooOoOoOO8O8OOo.8OOPd8O8O"&#x000A;              , ..:.::o:ooOoOOOO8OOOOo.FdO8O8"&#x000A;             , ..:.::o:ooOoOO8O888O8O,COCOO"&#x000A;            , . ..:.::o:ooOoOOOO8OOOOCOCO"&#x000A;             . ..:.::o:ooOoOoOO8O8OCCCC"o&#x000A;                . ..:.::o:ooooOoCoCCC"o:o&#x000A;                . ..:.::o:o:,cooooCo"oo:o:&#x000A;             `   . . ..:.:cocoooo"'o:o:::'&#x000A;             .`   . ..::ccccoc"'o:o:o:::'&#x000A;            :.:.    ,c:cccc"':.:.:.:.:.'&#x000A;          ..:.:"'`::::c:"'..:.:.:.:.:.'  http://www.chris.com/ASCII/&#x000A;        ...:.'.:.::::"'    . . . . .'&#x000A;       .. . ....:."' `   .  . . ''&#x000A;     . . . ...."'&#x000A;     .. . ."'     -hrr-&#x000A;    .&#x000A;&#x000A;&#x000A;                                              It's a planet!&#x000A;%strong This shouldn't be bold!</pre>
-<strong>This should!</strong>
-<textarea>      ___           ___           ___           ___ &#x000A;     /\__\         /\  \         /\__\         /\__\&#x000A;    /:/  /        /::\  \       /::|  |       /:/  /&#x000A;   /:/__/        /:/\:\  \     /:|:|  |      /:/  / &#x000A;  /::\  \ ___   /::\~\:\  \   /:/|:|__|__   /:/  /  &#x000A; /:/\:\  /\__\ /:/\:\ \:\__\ /:/ |::::\__\ /:/__/   &#x000A; \/__\:\/:/  / \/__\:\/:/  / \/__/~~/:/  / \:\  \   &#x000A;      \::/  /       \::/  /        /:/  /   \:\  \  &#x000A;      /:/  /        /:/  /        /:/  /     \:\  \ &#x000A;     /:/  /        /:/  /        /:/  /       \:\__\&#x000A;     \/__/         \/__/         \/__/         \/__/&#x000A;     &#x000A;     Many&#x000A;                   thanks&#x000A;           to&#x000A;                                http://www.network-science.de/ascii/
-<strong>indeed!</strong></textarea>
-</div>
-<div class='foo'>
-13
-</div>
-<pre>       __     ______        __               ______&#x000A;.----.|  |--.|__    |.----.|  |--..--------.|  __  |&#x000A;|  __||     ||__    ||  __||    < |        ||  __  |&#x000A;|____||__|__||______||____||__|__||__|__|__||______|</pre>
-<pre>foo
-bar</pre>
