--- test/templates/nuke_outer_whitespace.haml
@@ -1,144 +1,0 @@
-%p
-  %p
-    %q>
-      Foo
-%p
-  %p
-    %q{:a => 1 + 1}>
-      Foo
-%p
-  %p
-    %q> Foo
-%p
-  %p
-    %q{:a => 1 + 1}> Foo
-%p
-  %p
-    %q>
-      = "Foo"
-%p
-  %p
-    %q{:a => 1 + 1}>
-      = "Foo"
-%p
-  %p
-    %q>= "Foo"
-%p
-  %p
-    %q{:a => 1 + 1}>= "Foo"
-%p
-  %p
-    %q>
-      = "Foo\nBar"
-%p
-  %p
-    %q{:a => 1 + 1}>
-      = "Foo\nBar"
-%p
-  %p
-    %q>= "Foo\nBar"
-%p
-  %p
-    %q{:a => 1 + 1}>= "Foo\nBar"
-%p
-  %p
-    - tab_up
-    foo
-    %q>
-      Foo
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q{:a => 1 + 1}>
-      Foo
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q> Foo
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q{:a => 1 + 1}> Foo
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q>
-      = "Foo"
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q{:a => 1 + 1}>
-      = "Foo"
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q>= "Foo"
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q{:a => 1 + 1}>= "Foo"
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q>
-      = "Foo\nBar"
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q{:a => 1 + 1}>
-      = "Foo\nBar"
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q>= "Foo\nBar"
-    bar
-    - tab_down
-%p
-  %p
-    - tab_up
-    foo
-    %q{:a => 1 + 1}>= "Foo\nBar"
-    bar
-    - tab_down
-%p
-  %p
-    %q>
-%p
-  %p
-    %q>/
-%p
-  %p
-    %q{:a => 1 + 1}>
-%p
-  %p
-    %q{:a => 1 + 1}>/
