--- test/templates/partial_layout.haml
@@ -1,3 +1,0 @@
-%h1 Partial layout used with for block:
-= render :layout => 'layout_for_partial' do
-  %p Some content within a layout
