--- test/templates/whitespace_handling.haml
@@ -1,87 +1,0 @@
-#whitespace_test
-  = test_partial "text_area", :value => "Oneline"
-  = test_partial "text_area", :value => "Two\nlines"
-  ~ test_partial "text_area", :value => "Oneline"
-  ~ test_partial "text_area", :value => "Two\nlines"
-  #flattened~ test_partial "text_area", :value => "Two\nlines"
-.hithere
-  ~ "Foo bar"
-  ~ "<pre>foo bar</pre>"
-  ~ "<pre>foo\nbar</pre>"
-  %p~ "<pre>foo\nbar</pre>"
-  %p~ "foo\nbar"
-.foo
-  ~ 13
-  ~ "<textarea>\na\n</textarea>".each_line do |l|
-    - haml_concat l.strip
-#whitespace_test
-  = test_partial "text_area", :value => "Oneline"
-  = test_partial "text_area", :value => "Two\nlines"
-  = find_and_preserve test_partial("text_area", :value => "Oneline")
-  = find_and_preserve test_partial("text_area", :value => "Two\nlines")
-  #flattened= find_and_preserve test_partial("text_area", :value => "Two\nlines")
-.hithere
-  = find_and_preserve("Foo bar")
-  = find_and_preserve("<pre>foo bar</pre>")
-  = find_and_preserve("<pre>foo\nbar</pre>")
-  %p= find_and_preserve("<pre>foo\nbar</pre>")
-  %p= find_and_preserve("foo\nbar")
-  %pre
-    :preserve
-                                                       ___
-                                                    ,o88888
-                                                 ,o8888888'
-                           ,:o:o:oooo.        ,8O88Pd8888"
-                       ,.::.::o:ooooOoOoO. ,oO8O8Pd888'"
-                     ,.:.::o:ooOoOoOO8O8OOo.8OOPd8O8O"
-                    , ..:.::o:ooOoOOOO8OOOOo.FdO8O8"
-                   , ..:.::o:ooOoOO8O888O8O,COCOO"
-                  , . ..:.::o:ooOoOOOO8OOOOCOCO"
-                   . ..:.::o:ooOoOoOO8O8OCCCC"o
-                      . ..:.::o:ooooOoCoCCC"o:o
-                      . ..:.::o:o:,cooooCo"oo:o:
-                   `   . . ..:.:cocoooo"'o:o:::'
-                   .`   . ..::ccccoc"'o:o:o:::'
-                  :.:.    ,c:cccc"':.:.:.:.:.'
-                ..:.:"'`::::c:"'..:.:.:.:.:.'  http://www.chris.com/ASCII/
-              ...:.'.:.::::"'    . . . . .'
-             .. . ....:."' `   .  . . ''
-           . . . ...."'
-           .. . ."'     -hrr-
-          .
-
-
-                                                    It's a planet!
-      %strong This shouldn't be bold!
-  %strong This should!
-  %textarea
-    :preserve
-            ___           ___           ___           ___ 
-           /\__\         /\  \         /\__\         /\__\
-          /:/  /        /::\  \       /::|  |       /:/  /
-         /:/__/        /:/\:\  \     /:|:|  |      /:/  / 
-        /::\  \ ___   /::\~\:\  \   /:/|:|__|__   /:/  /  
-       /:/\:\  /\__\ /:/\:\ \:\__\ /:/ |::::\__\ /:/__/   
-       \/__\:\/:/  / \/__\:\/:/  / \/__/~~/:/  / \:\  \   
-            \::/  /       \::/  /        /:/  /   \:\  \  
-            /:/  /        /:/  /        /:/  /     \:\  \ 
-           /:/  /        /:/  /        /:/  /       \:\__\
-           \/__/         \/__/         \/__/         \/__/
-           
-           Many
-                         thanks
-                 to
-                                      http://www.network-science.de/ascii/
-    %strong indeed!
-.foo
-  = find_and_preserve(13)
-%pre
-  :preserve
-           __     ______        __               ______
-    .----.|  |--.|__    |.----.|  |--..--------.|  __  |
-    |  __||     ||__    ||  __||    < |        ||  __  |
-    |____||__|__||______||____||__|__||__|__|__||______|
-%pre
-  :preserve
-    foo
-  bar
